refactor: move SelectedUtxosHeader component to shared location for reuse.
